Fix issue #63
Currently lbry will open another process even if it is already running.
This is because it tries to put a lockfile in /tmp, which doesn't work.

As recommended in the Flatpak documentation:

"If an application uses $TMPDIR to contain lock files you may want to add a
wrapper script that sets it to $XDG_RUNTIME_DIR/app/$FLATPAK_ID."